+++ lua/autolist/config.lua
@@ -60,6 +60,7 @@ local default_config = {
     fill = "x", -- if you do the above two customisations, your checkbox could be (x) instead of [x]
   },
   loose_lists = false, -- experimental: treat single blank lines between items as part of a list
+  content_indent = true, -- align child lists to parent content (e.g. 3 spaces for "1. ")
 
   -- this is all based on lua patterns, see "Defining custom lists" for a nice article to learn them
 }